feat: update winRate method in PlayerStats to return an integer percentage instead of a double

+++ src/main/java/com/villo/truco/profile/application/dto/PlayerStatsDTO.java
@@ -1,5 +1,5 @@
 package com.villo.truco.profile.application.dto;
 
-public record PlayerStatsDTO(int matchesPlayed, int matchesWon, int matchesLost, double winRate) {
+public record PlayerStatsDTO(int matchesPlayed, int matchesWon, int matchesLost, int winRate) {
 
 }
